refactor: reposicionamiento de carpetas contexts, services, hooks. #30
