+++ src/App.tsx
@@ -5,17 +5,16 @@ import {
   Navigate,
 } from "react-router-dom";
 import { motion, AnimatePresence } from "framer-motion";
-//PROVIDERS
-import { UserProvider } from "./user/contexts/UserProvider";
+// PROVIDERS
+import { UserProvider } from "./user/contexts/userContext/UserProvider";
 import { YearProvider } from "./admin/contexts/yearContext/YearProvider";
 import { TeacherProvider } from "./admin/contexts/teacherContext/TeacherProvider";
 import { CourseProvider } from "./admin/contexts/courseContext/CourseProvider";
 import { StudentProvider } from "./admin/contexts/studentContext/StudentProvider";
-
+// PAGES
+import AdminDashboard from "./admin/pages/Dashboard/AdminDashboardPage";
 import { useAuth } from "./user/hooks/useAuth";
 import Login from "./user/pages/Login/LoginPage";
-import AdminDashboard from "./admin/pages/Dashboard/AdminDashboardPage";
-import { AdminRoutes } from "./admin/routes";
 import styles from "./App.module.css";
 
 import OverviewView from "./admin/views/OverviewView/OverviewView";
