+++ src/admin/contexts/yearContext/YearContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import type { YearContextType } from "./YearContext.types";
+import type { YearContextType } from "./YearContext.type";
 
 export const YearContext = createContext<YearContextType | undefined>(
   undefined
