+++ src/admin/services/teacher/TeacherService.ts
@@ -1,18 +1,20 @@
 import api from "../../../shared/utils/api";
-import { urls } from "../../urls";
+import { urls } from "../urls";
 
 export interface CreateTeacherPayload {
-    name:string,
-    lastname: string,
-    dni: string,
-    email: string,
+  name: string;
+  lastname: string;
+  dni: string;
+  email: string;
 }
 
-const registerTeacherApi = async (data: CreateTeacherPayload): Promise<void> => {
+const registerTeacherApi = async (
+  data: CreateTeacherPayload
+): Promise<void> => {
   try {
     await api.post(urls.Teacher, data);
   } catch (error) {
-    console.error("Error al crear el docente:", error);
+    console.error("Error al crear el docente:", error); // TODO: REMOVE_DEBUG
     throw error;
   }
 };
