--- src/urls.ts
@@ -1,4 +1,0 @@
-export const urlBase = {
-    server: 'https://play2learn.backend.desarrollo.systemsbinary.com',
-    localhost: 'http://localhost:8080'
-}
