+++ src/user/contexts/userContext/UserContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import type { UserContextType } from "./UserContext.types";
+import type { UserContextType } from "./UserContext.type";
 
 export const UserContext = createContext<UserContextType | undefined>(
   undefined
