+++ src/user/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { UserContext } from "../contexts/UserContext";
+import { UserContext } from "../contexts/userContext/UserContext";
 
 export const useAuth = () => {
   const context = useContext(UserContext);
