+++ src/user/services/Login/LoginService.ts
@@ -11,7 +11,7 @@ const loginApi = async (data: LoginPayload): Promise<any> => {
     const response = await api.post(urls.login, data);
     return response.data;
   } catch (error) {
-    console.error("Error en LogIn:", error);
+    console.error("Error en LogIn:", error); // TODO: REMOVE_DEBUG
     throw error;
   }
 };
